refactor(web): move the projects provider into lib/projects
It is project code living in a providers/ grab-bag. The feature already owns
hooks, types, svc and components; the context belongs beside them.

Twenty-one importers repointed, and the test moves with it.
