+++ web/src/app/app/layout.tsx
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import type { Route } from "next";
 import { unstable_noStore as noStore } from "next/cache";
 import { requireAuth } from "@/lib/auth/svcSS";
-import { ProjectsProvider } from "@/providers/ProjectsContext";
+import { ProjectsProvider } from "@/lib/projects/providers";
 import { VoiceModeProvider } from "@/providers/VoiceModeProvider";
 import AppSidebar from "@/sections/sidebar/AppSidebar";
 import { RootLayout } from "@opal/layouts";
