+++ web/src/refresh-components/popovers/ActionsPopover/ActionLineItem.tsx
@@ -13,7 +13,7 @@ import { noProp } from "@/lib/utils";
 import { cn } from "@opal/utils";
 import type { IconProps } from "@opal/types";
 import { SvgChevronRight, SvgKey, SvgSettings, SvgSlash } from "@opal/icons";
-import { useProjectsContext } from "@/providers/ProjectsContext";
+import { useProjectsContext } from "@/lib/projects/providers";
 import { useRouter } from "next/navigation";
 import type { Route } from "next";
 import EnabledCount from "@/refresh-components/EnabledCount";
