+++ web/src/sections/modals/UserFilesModal.tsx
@@ -2,7 +2,7 @@
 
 import React, { useRef, useState, useEffect, useMemo } from "react";
 import { InputTypeIn } from "@opal/components";
-import { ProjectFile } from "@/providers/ProjectsContext";
+import { ProjectFile } from "@/lib/projects/providers";
 import Text from "@/refresh-components/texts/Text";
 import type { IconProps } from "@opal/types";
 import { getFileExtension, isImageExtension } from "@/lib/utils";
