+++ web/src/sections/sidebar/AppSidebar.tsx
@@ -45,7 +45,7 @@ import {
 } from "@/lib/projects/components";
 import CreateProjectModal from "@/sections/modals/CreateProjectModal";
 import { MoveCustomAgentChatModal } from "@/lib/agents/components";
-import { useProjectsContext } from "@/providers/ProjectsContext";
+import { useProjectsContext } from "@/lib/projects/providers";
 import { removeChatSessionFromProject } from "@/lib/projects/svc";
 import type { Project } from "@/lib/projects/types";
 import { SidebarLayouts, useSidebarState } from "@opal/layouts";
